Move Tuya listener classes to separate module
