+++ homeassistant/components/tuya/diagnostics.py
@@ -12,8 +12,8 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers import device_registry as dr, entity_registry as er
 from homeassistant.helpers.device_registry import DeviceEntry
 
-from . import TuyaConfigEntry
 from .const import DOMAIN, DPCode
+from .coordinator import TuyaConfigEntry
 
 _REDACTED_DPCODES = {
     DPCode.ALARM_MESSAGE,
