+++ homeassistant/components/tuya/scene.py
@@ -11,8 +11,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import TuyaConfigEntry
 from .const import DOMAIN
+from .coordinator import TuyaConfigEntry
 
 
 async def async_setup_entry(
